% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.databases-package.R
\docType{package}
\name{chem.databases-package}
\alias{chem.databases}
\alias{chem.databases-package}
\title{chem.databases: Collection of 3 Chemical Databases}
\description{
chem.databases provides the Multi-Species Acute Toxicity Database (CAS &
SMILES columns only) combined with the Toxic Substances Control Act (TSCA)
Inventory and the Agency for Toxic Substances and Disease Registry (ATSDR)
Database in 2 data sets. One data set has a focus on the latter 2 databases
and one data set focuses on the former database. Also contains the
collection of chemical data from Wikipedia compiled in the US EPA CompTox
Chemicals Dashboard.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/iembry/chem.databases}
  \item Report bugs at \url{https://gitlab.com/iembry/chem.databases/-/issues}
}

}
\author{
\strong{Maintainer}: Irucka Embry \email{iembry@ecoccs.com}

}
\keyword{internal}
