% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_wiki.R
\docType{data}
\name{chem_wiki}
\alias{chem_wiki}
\title{CompTox Chemicals Dashboard From Wikipedia}
\format{
A data.table data frame with 19,239 rows and 9 variables:
\describe{
item{CAS}{Chemical Abstracts Service (CAS) Registry Number}
item{Substance Name}{Preferred Chemical Substance Name}
item{IUPAC Name}{IUPAC Chemical Name}
item{Molecular Formula}{Chemical Molecular Formula}
item{SMILES}{Simplified Molecular-Input Line-Entry System (SMILES) Chemical Structural Notation}
item{InChI}{International Chemical Identifier (InChI) Chemical Structural Notation}
item{InChIKey}{Hashed Version of the InChI}
item{Average Mass}{Average Chemical Molecular Mass}
item{Monoisotopic Mass}{Single Chemical Isotope Mass}
}
}
\source{
United States Environmental Protection Agency (US EPA) / Wikimedia Foundation, Inc. "CompTox Chemicals Dashboard v2.2.1", \url{https://comptox.epa.gov/dashboard/chemical-lists/WIKIPEDIA}.
}
\usage{
chem_wiki
}
\description{
A table containing chemical data from the CompTox Chemicals Dashboard which
is derived from various Wikipedia pages.
}
\keyword{datasets}
